Fix trace allocation header ordering
Current bug: NETDATA_TRACE_ALLOCATIONS builds fail in the mmap trace path because nd-mallocz.h embeds avl_t before libnetdata.h includes avl.h, and nd-mmap.c calls malloc_trace_mmap() / malloc_trace_munmap() without visible prototypes. Potential bug: the debug trace allocation build path cannot validate mmap trace accounting even though default builds compile with the trace branch disabled.

Root cause: libnetdata.h includes memory/nd-mallocz.h before avl/avl.h, while full avl.h cannot be pulled earlier because it also defines lock-dependent AVL tree types. The mmap trace helpers are defined in nd-mallocz.c but were not declared by the trace allocation header.

Solution: move the existing avl_t and avl_t_packed typedefs into a dependency-free avl-types.h leaf header, include it from avl.h and the trace-only nd-mallocz.h block, declare malloc_trace_mmap() and malloc_trace_munmap(), and list the new header in CMake.

Functional impact: no executable code changes and no default-build behavior changes. Trace builds keep the same struct layouts, trace counters, mmap/munmap logic, and helper implementations; the changed surface is type/prototype visibility only.

Performance impact: none. The patch adds no runtime branches, allocations, locks, atomics, or extra work.

Caller and contract impact: avl.h continues to expose the same AVL typedefs and full API. nd_mmap(), nd_munmap(), nd_mmap_advanced(), and pulse trace allocation callers keep the same return, ownership, fd, counter, and traversal contracts.

(cherry picked from commit 0ba5281bb402f4a0ba692f38408c349e26a3dc89)
